Update method for setting Alipay certificates in client.go
Modified the instructions for setting Alipay public key certificates in the comments of client.go file. Instead of using `client.SetAlipayRootCertSN()` and `client.SetAppCertSN()`, it's recommended to use `client.SetCertSnByContent()` or `client.SetCertSnByPath()`. This change was made to enhance the clarity of the process for setting up Alipay public key certificates.